requirements.txt
robotframework>=4.1.2
jmespath>=1.0.1
python-dateutil>=2.8.2
requests>=2.31.0
thefuzz>=0.20.0
pyyaml>=6.0.1
rw-cli-keywords>=0.0.23
azure-cosmos>=4.5.1